# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: Bank-Info-API
  description: OpenCode practice 2023
  version: 0.0.3
servers:
  - url: http://localhost:8081/

# security:
#   - basicAuth: []
paths:
...
  /info:
      get:
        tags:
          - Bank Info
        description: Get all bank info. Default sorting ASC by id.
        operationId: getAllInfo
        summary: getAllInfo
        parameters: 
          - $ref: '#/components/parameters/SizeQuery'
          - $ref: '#/components/parameters/PageQuery'
          - $ref: '#/components/parameters/InfoNameQuery'
          - $ref: '#/components/parameters/DeletedQuery'
        responses:
          200:
            $ref: '#/components/responses/InfoPaginate'
            
      post:
        tags:
          - Bank Info
        description: Create bank info
        operationId: createInfo
        summary: createInfo
        requestBody:
          $ref: '#/components/requestBodies/CreateInfo'
        responses:
          201:
            $ref: '#/components/responses/201'
          400:
            $ref: '#/components/responses/InputValidationError'
  /info/{id}:
      put:
        tags:
          - Bank Info
        description: Update bank info by id
        operationId: updateInfoById
        summary: updateInfoById
        parameters:
          - $ref: '#/components/parameters/InfoIdPath'
        requestBody:
          $ref: '#/components/requestBodies/CreateInfo'
        responses:
          204:
            $ref: '#/components/responses/204'
          400:
            $ref: '#/components/responses/InputValidationError'
          404:
            $ref: '#/components/responses/InfoNotFoundError'
      
      delete:
        tags:
          - Bank Info
        description: Delete bank info by id
        operationId: deleteInfoById
        summary: deleteInfoById
        parameters:
          - $ref: '#/components/parameters/InfoIdPath'
        responses:
          204:
            $ref: '#/components/responses/204'
          404:
            $ref: '#/components/responses/InfoNotFoundError'
            
      patch:
        tags:
          - Bank Info
        description: Restore bank info by id
        operationId: restoreInfoById
        summary: restoreInfoById
        parameters:
          - $ref: '#/components/parameters/InfoIdPath'
        responses:
          204:
            $ref: '#/components/responses/204'
          404:
            $ref: '#/components/responses/InfoNotFoundError'
  /info/{id}/manuals:
      get:
        tags:
          - Manuals
        description: Get manuals by bank info id. Default sorting ASC by id.
        operationId: getManualsByInfoId
        summary: getManualsByInfoId
        parameters: 
          - $ref: '#/components/parameters/InfoIdPath'
          - $ref: '#/components/parameters/SizeQuery'
          - $ref: '#/components/parameters/PageQuery'
          - $ref: '#/components/parameters/ManualCodeQuery'
          - $ref: '#/components/parameters/ManualDescriptionQuery'
          - $ref: '#/components/parameters/DeletedQuery'
        responses:
          200:
            $ref: '#/components/responses/ManualPaginate'
          404:
            $ref: '#/components/responses/InfoNotFoundError'
            
      post:
        tags:
          - Manuals
        description: Add manual record(row) by info id
        operationId: addManual
        summary: addManual
        parameters: 
          - $ref: '#/components/parameters/InfoIdPath'
        requestBody:
          $ref: '#/components/requestBodies/CreateManual'
        responses:
          201:
            $ref: '#/components/responses/201'
          400:
            $ref: '#/components/responses/InputValidationError'
          404:
            $ref: '#/components/responses/InfoNotFoundError'
  /info/manuals/{id}:
      put:
        tags:
          - Manuals
        description: Update manual by id
        operationId: updateManualById
        summary: updateManualById
        parameters:
          - $ref: '#/components/parameters/ManualIdPath'
        requestBody:
          $ref: '#/components/requestBodies/CreateManual'
        responses:
          204:
            $ref: '#/components/responses/204'
          400:
            $ref: '#/components/responses/InputValidationError'
          404:
            $ref: '#/components/responses/ManualNotFoundError'
      
      delete:
        tags:
          - Manuals
        description: Delete manual by id
        operationId: deleteManualById
        summary: deleteManualById
        parameters:
          - $ref: '#/components/parameters/ManualIdPath'
        responses:
          204:
            $ref: '#/components/responses/204'
          404:
            $ref: '#/components/responses/ManualNotFoundError'
            
      patch:
        tags:
          - Manuals
        description: Restore manual by id
        operationId: restoreManualById
        summary: restoreManualById
        parameters:
          - $ref: '#/components/parameters/ManualIdPath'
        responses:
          204:
            $ref: '#/components/responses/204'
          404:
            $ref: '#/components/responses/ManualNotFoundError'
components:
...
  parameters:
    SizeQuery:
      name: size
      in: query
      description: Count of items in one page
      required: true
      schema:
        type: integer
        default: 10
    PageQuery:
      name: page
      in: query
      description: Number of page items
      required: true
      schema:
        type: integer
        default: 1
...
    InfoNameQuery:
      name: name
      in: query
      description: Filter by bank info name
      required: false
      schema:
        type: string
    ManualCodeQuery:
      name: code
      in: query
      description: Filter by bank info code
      required: false
      schema:
        type: string
    ManualDescriptionQuery:
      name: description
      in: query
      description: Filter by bank info description
      required: false
      schema:
        type: string
    DeletedQuery:
      name: deleted
      in: query
      description: If this filter true - search area include deleted items
      required: false
      schema:
        type: boolean
        default: false
    InfoIdPath:
      name: id
      in: path
      description: Bank info id
      required: true
      schema:
        type: integer
    ManualIdPath:
      name: id
      in: path
      description: Manual id
      required: true
      schema:
        type: integer
  requestBodies:
...
    CreateManual:
      description: Manual info
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateManual'
    CreateInfo:
      description: Bank info
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateInfo'
...
  responses:
    ManualPaginate:
      description: All Bank Info
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManualPaginate"
    InfoPaginate:
      description: All Bank Info
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InfoPaginate"
...
    InfoNotFoundError:
      description: Bank info not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ManualNotFoundError:
      description: Manual not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InputValidationError:
      description: Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    201:
      description: Created
    204:
      description: No content
  schemas:
...
    AuditFields:
      type: object
      properties:
        isDeleted:
          type: boolean
        # createdBy:
        #   type: integer
        createDateTime:
          type: string
          format: date-time
        # changedBy:
        #   type: integer
        changeDateTime:
          type: string
          format: date-time
...
    CreateManual:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: RQST
        description: Ответ на ЭС-запрос участника
    Manual:
      type: object
      properties:
        id:
          type: integer
      allOf:
        - $ref: '#/components/schemas/CreateManual'
        - $ref: '#/components/schemas/AuditFields'
      example:
        id: 0
        code: RQST
        description: Ответ на ЭС-запрос участника
        isDeleted: false
        createDateTime: 2023-05-17T18:31:47Z
        changeDateTime: 2023-05-17T18:31:47Z
    Manuals:
      type: array
      items:
        $ref: '#/components/schemas/Manual'
    CreateInfo:
      type: object
      properties:
        name:
          type: string
      example:
        name: Причина создания ЭС
    Info:
      type: object
      properties:
        id:
          type: integer
      allOf:
        - $ref: '#/components/schemas/CreateInfo'
        - $ref: '#/components/schemas/AuditFields'
      example:
        id: 0
        name: Причина создания ЭС
        isDeleted: false
        createDateTime: 0
        changeDateTime: 0
    ConfigPagination:
      type: object
      properties:
        countOfPages:
          type: integer
        countOfItems:
          type: integer
      example:
        countOfPages: 5
        countOfItems: 50
    InfoPaginate:
      type: object
      properties:
        items:
          $ref: '#/components/schemas/InfoArray'
        config:
          $ref: '#/components/schemas/ConfigPagination'
    ManualPaginate:
      type: object
      properties:
        items:
          $ref: '#/components/schemas/Manuals'
        config:
          $ref: '#/components/schemas/ConfigPagination'
    InfoArray:
      type: array
      items:
        $ref: '#/components/schemas/Info'
    Error:
      type: object
      properties:
        message:
          type: string
          nullable: false
      example:
        message: error.message